% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgBase.R
\name{cgBase}
\alias{cgBase}
\title{create CGDS object}
\usage{
cgBase(address = "https://www.cbioportal.org/")
}
\arguments{
\item{address}{string, web address}
}
\value{
cdgs object, and prints the first 2 columns of the cdgs object (the first column contains the IDs, to be used in later functions)
}
\description{
Imports:
cgdsr
}
\details{
uses cgdsr::CGDS(address)
}
\examples{
cgds <- cgBbase()
}
