% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgData.R
\name{cgData}
\alias{cgData}
\title{output data of the specified type}
\usage{
cgData(
  cgds,
  cancer_study_id,
  profile_name,
  caselist_name = "All samples",
  genes = NA
)
}
\arguments{
\item{cgds}{object, output from the cgBase() function}

\item{cancer_study_id}{string, the ID as given in the first column of the cgBase() printout}

\item{profile_name}{string, profile name (not ID) to be used (obtained from the printout of the cgStudy() function)}

\item{caselist_name}{string, caselist name (not ID) to be used (obtained from the printout of the cgStudy() function)}

\item{genes}{string vector with genes of interest. Symbols, EntrezIds and EnsemblIds work. If nothing (NA) is provided, all genes in the GO database will be used. This will take some time, so if you only need a specific set of genes, specify it here.}
}
\value{
data.frame, each row is a patient, each column is a gene
}
\description{
Imports:
cgdsr,
org.Hs.eg.db
}
\examples{
cgds <- cgBbase()
mystudy <- "laml_tcga"
cgStudy(cgds, mystudy)
myData <- cgData(cgds, mystudy, profile_name="mRNA expression (RNA Seq RPKM)", genes=c("FLT3","TP53"))
}
