% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgInfo.R
\name{cgInfo}
\alias{cgInfo}
\title{get info about a study}
\usage{
cgInfo(cgds, cancer_study_id)
}
\arguments{
\item{cgds}{object, output from the cgBase() function}

\item{cancer_study_id}{string, the ID as given in the first column of the cgBase() printout}
}
\value{
nothing, but prints out two tables with information for the follow-up function
}
\description{
get info about a study
}
\details{
uses cgdsr::getCancerStudies(cgds)
}
\examples{
cgds <- cgBbase()
mystudy <- "laml_tcga"
cgInfo(cgds, mystudy)
}
