% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgMutation.R
\name{cgMutation}
\alias{cgMutation}
\title{convenient shortcut to find mutations}
\usage{
cgMutation(
  cgds,
  cancer_study_id,
  genes = NA,
  caselist_name = "Samples with mutation data"
)
}
\arguments{
\item{cgds}{object, output from the cgBase() function}

\item{cancer_study_id}{string, the as given in the cgBase() printout}

\item{genes}{genes string vector with genes of interest. Symbols, EntrezIds and EnsemblIds work. If nothing (NA) is provided, all genes in the GO database will be used. This will take some time, so if you only need a specific set of genes, specify it here.}

\item{caselist_name}{string, should be chosen so that only samples with mutation data are used.}
}
\value{
data.frame, each row is a patient, each column is a gene
}
\description{
convenient shortcut to find mutations
}
\examples{
cgds <- cgBbase()
mystudy <- "laml_tcga"
myMutations <- cgData(cgds, mystudy, genes=c("FLT3","TP53"))
}
