% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiNetplot.R
\name{DiNetplot}
\alias{DiNetplot}
\title{DiNetplot}
\usage{
DiNetplot(cgNetwork_list)
}
\arguments{
\item{cgNetwork_list}{list of igraph objects, generated by cgNetwork().}
}
\description{
Displays an interactive network plot.
}
\examples{
example.file <- system.file("extdata", "example.RData", package="cgNetwork")
load(example.file)
DiNetplot(cgNetwork_list)
}
