% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_info.R
\name{mutation_info}
\alias{mutation_info}
\title{mutation_info Function}
\usage{
mutation_info(study_name = "STAD", gene = "TP53", pipeline = "metect2")
}
\arguments{
\item{study_name}{name of TCGA study. Defaults to "STAD" (Stomach adenocarcinoma). Find more TCGA Study Abbreviations: https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations}

\item{gene}{Hugo Symbol of gene Default to "TP53"}

\item{pipeline}{Variant calling is performed using several separate pipelines; muse, varscan2, somaticsniper, mutect2. Default to mutect2. Find details: https://docs.gdc.cancer.gov/Data/Bioinformatics_Pipelines/DNA_Seq_Variant_Calling_Pipeline/#somatic-variant-calling-workflow}
}
\description{
This function allows the user to download mutation infomation.
}
\examples{
mut_info <- mutation_info("STAD", "TP53")
mut_info <- mutation_info("PAAD", "KRAS")
}
