% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_finder.R
\name{neighbor_finder}
\alias{neighbor_finder}
\title{neighbor_finder}
\usage{
neighbor_finder(
  big_cor_matrix,
  gene = "CDKN1A",
  cor_method = "spearman",
  cor.cut.off = 0.39,
  weight.cut.off = 0.5
)
}
\arguments{
\item{big_cor_matrix}{bigmatrix (bigmemory package) of Spearman's correlation coefficient from TCGA RNAseq countdata}

\item{gene}{Interesting gene to find network. Defaults to "CDKN1A".}

\item{cor_method}{correlation coeffiecint method "spearman" or "pearson". Defaults to "spearman"}

\item{cor.cut.off}{cut off value of correlation coeffiecint to build network. Defaults to .39}

\item{weight.cut.off}{cut off value of network weight to reduce edges. Defaults to .5}
}
\description{
This function allows the user to find neighbor genes network of interested gene.
}
\examples{
common_neighbor <- neighbor_finder(big_cor_matrix, gene="CDKN1A", cor_method = "spearman", cor.cut.off=.39, weight.cut.off=.5)
}
