% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_RNAseq_RSEM.R
\name{TCGA_RNAseq_RSEM}
\alias{TCGA_RNAseq_RSEM}
\title{TCGA_RNAseq_RSEM Function}
\usage{
TCGA_RNAseq_RSEM(study_name = "STAD")
}
\arguments{
\item{study_name}{name of TCGA study. Defaults to STAD (Stomach adenocarcinoma). Find more TCGA Study Abbreviations: https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations}
}
\description{
This function allows the user to download TCGA RNAseq data.
}
\examples{
countdata <- TCGA_RNAseq_RSEM("STAD")
countdata <- TCGA_RNAseq_RSEM("PAAD")
}
