% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGANetwork.R
\name{TCGANetwork}
\alias{TCGANetwork}
\title{TCGANetwork}
\usage{
TCGANetwork(
  countdata,
  mut_df,
  common_neighbor,
  cor_method = "spearman",
  weight.cut.off = 0.5
)
}
\arguments{
\item{countdata}{TCGA RNAseq countdata generated by TCGA_RNAseq_RSEM().}

\item{mut_df}{Mutation status generated by mutation_info().}

\item{common_neighbor}{Neighbor genes of interesting gene generated by neighbor_finder().}

\item{cor_method}{correlation coeffiecint method "spearman" or "pearson". Defaults to "spearman"}

\item{weight.cut.off}{cut off value of network weight to reduce edges. Defaults to .5}
}
\description{
This function allows the user to find neighbor genes network of mutation status of interested gene.
It returns two igraph lists.
}
\examples{
TCGANetwork_list <- TCGANetwork(countdata, mut_df, common_neighbor, cor_method = "spearman", weight.cut.off=.5)
}
