% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgNetwork.R
\name{cgNetwork}
\alias{cgNetwork}
\title{cgNetwork}
\usage{
cgNetwork(
  countdata,
  mut_df,
  common_neighbor,
  cor_method = "spearman",
  weight.cut.off = 0.5
)
}
\arguments{
\item{mut_df}{data.frame, output of the cgMutation or mutation_info function.}

\item{common_neighbor}{character vector, names of genes considered neighbors of the gene of interest. Output of the neighbor_finder function.}

\item{cor_method}{string, correlation coeffiecint method "spearman" or "pearson". Defaults to "spearman".}

\item{weight.cut.off}{numeric, cut off value of network weight to reduce edges. Defaults to 0.5.}

\item{data.frame, }{output of the cgData or TCGA_RNAseq_RSEM function.}
}
\value{
list of two igraph objects
}
\description{
cgNetwork
}
\examples{
cgNetwork_list <- cgNetwork(countdata, mut_df, common_neighbor, cor_method = "spearman", weight.cut.off=.5)
}
