% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiNetplot.R
\name{DiNetplot}
\alias{DiNetplot}
\title{DiNetplot}
\usage{
DiNetplot(cgNetwork_list)
}
\arguments{
\item{cgNetwork_list}{list of igraph objects, generated by cgNetwork().}

\item{interest_gene}{string, gene of interest same as used in the neighbor_finder function.}

\item{mut_gene}{string, gene by which patients should be separated. There will be one network for WT and one for mutated.}
}
\description{
Imports:
shiny,
igraph,
dplyr,
visNetwork
}
\details{
Displays an interactive network plot.
}
\examples{
DiNetplot(cgNetwork_list)
}
