% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiNetplot.R
\name{DiNetplot}
\alias{DiNetplot}
\title{DiNetplot}
\usage{
DiNetplot(TCGANetwork_list)
}
\arguments{
\item{TCGANetwork_list}{igraph list generated by TCGANetwork().}
}
\description{
This function display network plot using Shiny and visNetwork packages.
}
\examples{
DiNetplot(TCGANetwork_list)
}
