% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiNetplot.R
\name{DiNetplot}
\alias{DiNetplot}
\title{DiNetplot}
\usage{
DiNetplot(TCGANetwork_list)
}
\arguments{
\item{TCGANetwork_list}{igraph list generated by TCGANetwork().}

\item{interest_gene}{The gene that used in neighbor_finder function.}

\item{mut_gene}{The gene that used in mutation_info function.}
}
\description{
This function display network plot using Shiny and visNetwork packages.
}
\examples{
DiNetplot(TCGANetwork_list)
}
