% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_finder.R
\name{neighbor_finder}
\alias{neighbor_finder}
\title{neighbor_finder}
\usage{
neighbor_finder(
  countdata,
  gene = gene_of_interest,
  cor.cut.off = 0.39,
  weight.cut.off = 0.5
)
}
\arguments{
\item{countdata}{RNAseq countdata}

\item{gene}{string, gene of interest to find a network for.}

\item{cor.cut.off}{cut off value of correlation coeffiecint to build network. Defaults to .39}

\item{weight.cut.off}{cut off value of network weight to reduce edges. Defaults to .5}

\item{t}{Matrix Transpose}
}
\value{
character vector, genes that are considered neighbors of the gene of interest (including the gene of interest itself).
}
\description{
neighbor_finder
}
\examples{
common_neighbor <- neighbor_finder(big_cor_matrix, gene="CDKN1A", cor_method = "spearman", cor.cut.off=.39, weight.cut.off=.5, t=TRUE)
}
